+++ Resources/sql/update200.sql
@@ -1,0 +1,2 @@
+DELETE FROM `s_crontab` WHERE `action` = 'Shopware_Components_CronJob_FfuenfCronMonitoring';
+DELETE FROM `s_core_config_elements` WHERE `label` = 'Whitelist for CSP-Reporting');