+++ Subscriber/Cronjob.php
@@ -6,7 +6,7 @@
  * @category   Shopware
  * @package    Shopware\Plugins\FfuenfCommon
  * @author     Achim Rosenhagen / ffuenf - Pra & Rosenhagen GbR
- * @copyright  Copyright (c) 2020, Achim Rosenhagen / ffuenf - Pra & Rosenhagen GbR (https://www.ffuenf.de)
+ * @copyright  Copyright (c) 2021, Achim Rosenhagen / ffuenf - Pra & Rosenhagen GbR (https://www.ffuenf.de)
  *
  */
 
@@ -22,8 +22,7 @@ class Cronjob extends AbstractService implements SubscriberInterface
     public static function getSubscribedEvents()
     {
         return [
-            'Shopware_Components_CronJob_FfuenfCacheWarmUpCron' => 'onRunWarmUpCache',
-            'Shopware_Components_CronJob_FfuenfCronMonitoring'  => 'onCronMonitoring'
+            'Shopware_Components_CronJob_FfuenfCacheWarmUpCron' => 'onRunWarmUpCache'
         ];
     }
 
@@ -43,24 +42,6 @@ class Cronjob extends AbstractService implements SubscriberInterface
             $time_end = microtime(true);
             $time = $time_end - $time_start;
             return 'Cache wurde erfolgreich aufgewärmt!' . PHP_EOL . 'Dauer: ' . round($time, 0) . 'Sekunden';
-        } catch (\Exception $e) {
-            return $e->getMessage();
-        }
-    }
-
-    /**
-     * @param Enlight_Event_EventArgs $args
-     * @return string
-     */
-    public function onCronMonitoring(Enlight_Event_EventArgs $args)
-    {
-        try {
-            $time_start = microtime(true);
-            $command = Shopware()->DocPath() . 'bin/console ffuenf:cronmonitoring:check';
-            exec($command, $retArr, $retVal);
-            $time_end = microtime(true);
-            $time = $time_end - $time_start;
-            return 'CronMonitoring erfolgreich ausgeführt!' . PHP_EOL . 'Dauer: ' . round($time, 0) . 'Sekunden';
         } catch (\Exception $e) {
             return $e->getMessage();
         }